+++ gipit-back/app/api/post_sales_activities/[id]/route.ts
@@ -1,5 +1,6 @@
 import { PrismaClient } from '@prisma/client';
 import { NextRequest, NextResponse } from 'next/server';
+import { verifyToken } from '../../middleware'; 
 
 const prisma = new PrismaClient();
 
@@ -29,6 +30,12 @@ export async function GET(req: NextRequest, { params }: { params: { id: string }
   const { id } = params;
 
   try {
+
+    const verificationResult = verifyToken(req);
+    if (!verificationResult.valid) {
+      return NextResponse.json({ error: verificationResult.error }, { status: 403 });
+    }
+
     const postSalesActivity = await prisma.post_sales_activities.findUnique({
       where: { id: parseInt(id) },
     });
@@ -39,7 +46,7 @@ export async function GET(req: NextRequest, { params }: { params: { id: string }
 
     return NextResponse.json(postSalesActivity);
   } catch (error) {
-    return NextResponse.json({ error: 'Error obteniendo la actividad de ventas' }, { status: 500 });
+    return NextResponse.json({ error: `Error - ${error}` }, { status: 500 });
   }
 }
 
@@ -71,6 +78,12 @@ export async function PUT(req: NextRequest, { params }: { params: { id: string }
   const { candidate_management_id, benefit, description, date, associated_cost } = await req.json();
 
   try {
+
+    const verificationResult = verifyToken(req);
+    if (!verificationResult.valid) {
+      return NextResponse.json({ error: verificationResult.error }, { status: 403 });
+    }
+
     const updatedPostSalesActivity = await prisma.post_sales_activities.update({
       where: { id: parseInt(id) },
       data: {
@@ -85,7 +98,7 @@ export async function PUT(req: NextRequest, { params }: { params: { id: string }
 
     return NextResponse.json(updatedPostSalesActivity);
   } catch (error) {
-    return NextResponse.json({ error: 'Error actualizando la actividad de ventas' }, { status: 500 });
+    return NextResponse.json({ error: `Error - ${error}` }, { status: 500 });
   }
 }
 
@@ -110,12 +123,18 @@ export async function DELETE(req: NextRequest, { params }: { params: { id: strin
   const { id } = params;
 
   try {
+
+    const verificationResult = verifyToken(req);
+    if (!verificationResult.valid) {
+      return NextResponse.json({ error: verificationResult.error }, { status: 403 });
+    }
+
     await prisma.post_sales_activities.delete({
       where: { id: parseInt(id) },
     });
 
     return NextResponse.json({ message: 'Actividad de ventas eliminada con éxito' });
   } catch (error) {
-    return NextResponse.json({ error: 'Error eliminando la actividad de ventas' }, { status: 500 });
+    return NextResponse.json({ error: `Error - ${error}` }, { status: 500 });
   }
 }
